+++ scripts/step_rule/step_rule.js
@@ -103,9 +103,9 @@ function save() {
       let no = $(this).data('no');
       let id = $(this).data('id');
       let label = $('#label-'+no).val().trim();
-      let stepQty = parseDefault(parseInt($('#step-qty-'+no).val()), 0);
-      let limitQty = parseDefault(parseInt($('#limit-qty-'+no).val()), 0);
-      let freeQty = parseDefault(parseInt($('#free-qty-'+no).val()), 0);
+      let stepQty = parseDefault(parseFloat($('#step-qty-'+no).val()), 0);
+      let limitQty = parseDefault(parseFloat($('#limit-qty-'+no).val()), 0);
+      let freeQty = parseDefault(parseFloat($('#free-qty-'+no).val()), 0);
       let active = $('#active-'+no).val();
       let force = $('#force-'+no).is(':checked') ? 1 : 0;
       let highlight = $('#highlight-'+no).is(':checked') ? 1 : 0;
